Page-specific meta descriptions
This brings in the meta partial that the theme of this blog uses and
customizes it to set the meta description content to be the value
defined in the frontmatter of a post, if it exists. Otherwise, it uses
the site-wide description.

This also applies that description frontmatter to all blog post/article
type entries.