Add Canonical URL
This adds in metadata to the meta partial for the canonical link to the
post. It expects an additional piece of front-matter, named
`canonicalURL`, to be set should the canonical URL be somewhere else.

If the front-matter isn't specified, then the canonical URL is the
current full path.

All posts from the Gnar blog are updated to source their site as the
canonical URL.