Adding Pagefind
Inspired by [Justin Searls](https://justin.searls.co/links/2023-11-29-shout-out-to-pagefind-static-search/),
this adds [Pagefind](https://pagefind.app/) to be able to search the
posts on this site.

Running:
```
npx -y pagefind --site public
```

Running locally:
```
npx -y pagefind --site public --serve
```

To get this displayed, this creates a search partial that is exposed
through a shortcode. That shortcode is displayed on a new search page
that is provided on the top nav of the site.